requirements.txt
requests==2.31.0
google-generativeai==0.2.0
pymongo==4.5.0
ndjson==0.3.1
python-dotenv==1.0.0
